Refactor commom to core name
