+++ app/src/main/java/com/oskr19/easyshop/domain/core/failure/Failure.kt
@@ -1,4 +1,4 @@
-package com.oskr19.easyshop.domain.common.failure
+package com.oskr19.easyshop.domain.core.failure
 
 /**
  * Created by oscar.vergara on 24/07/2021
