+++ app/src/main/java/com/oskr19/easyshop/domain/core/network/INetworkHandler.kt
@@ -1,4 +1,4 @@
-package com.oskr19.easyshop.domain.common.network
+package com.oskr19.easyshop.domain.core.network
 
 /**
  * Created by oscar.vergara on 24/07/2021
