+++ app/src/main/java/com/oskr19/easyshop/presentation/core/UiEvent.kt
@@ -1,4 +1,4 @@
-package com.oskr19.easyshop.presentation.common
+package com.oskr19.easyshop.presentation.core
 
 /**
  * Created by oscar.vergara on 24/07/2021
