+++ app/src/main/java/com/oskr19/easyshop/presentation/core/viewmodel/BaseViewModel.kt
@@ -1,11 +1,11 @@
-package com.oskr19.easyshop.presentation.common.viewmodel
+package com.oskr19.easyshop.presentation.core.viewmodel
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import com.oskr19.easyshop.domain.common.failure.Failure
-import com.oskr19.easyshop.presentation.common.UiEvent
+import com.oskr19.easyshop.domain.core.failure.Failure
+import com.oskr19.easyshop.presentation.core.UiEvent
 
 /**
  * Created by oscar.vergara on 24/07/2021
